+++ .gitignore
@@ -1,0 +1,2 @@
+reveal.js/*
+node_modules/*
