+++ slides/misc.md
@@ -1,22 +1,18 @@
-# Subjects: 
+# Algorithms and model definitions
 
-- interpolation
-- optimization
-- differentiation
-
-- libraries:
-    - LinearMaps
-
+----
 
 # Linear Time iteration
 
 - $s_t \in R^{n_s} $: continuous states 
 - $x_t \in R^{n_x}$: continuous controls 
 - $\epsilon_t \in R^{n_e}$: exogenous shock 
-- state transition
+- State transition
 $$s_{t+1} = g(s_{t}, x_{t}, \epsilon_t)$$
 - Euler equation (aka arbitrage)
 $$E_t [ f(s_t, x_t, s_{t+1}, x_{t+1}) ] = 0$$
+
+----
 
 # Markov Decision Process with Continous Controls
 
@@ -58,7 +54,10 @@ Each function $\varphi^i$ is still defined by an infinite number of parameters.
 - characterize the values at a finite number of points...
 - ...and use interpolation to find the values between the grid points
 
-Our task:
+----
+
+## Setting up the state-space
+
 - define a discrete subset $\mathcal{C^d}$ of $\mathcal{C}$, the *grid*
     - cf CGrid, SGrid, $\times$
 - represent vectors of control values on the grid
@@ -99,6 +98,26 @@ $$F(\overrightarrow{x_t}, \overrightarrow{x_{t+1}}) \in \left(R^{n_x}\right)^{|\
 
 ----
 
+##  Measuring optimality (complementarities)
+
+- In some cases the optimality condition is met only if a given constraint is not met.
+- In that case we can rewrite the optimality condition as:
+
+$$E_t[ f(s_t, x_t, s_{t+1}, x_{t+1})] \geq 0 \perp b(x_t)\geq 0$$
+
+where $b(x_t)$ is the constraint.
+- By convention $a\perp b$ exactly means $a\geq0$,$b\geq0$ and $a b = 0$
+
+
+- It often arises from constrained optimization
+- In practice, one can either:
+    - resort to a solver which can deal with complementarities (like `NLsolve`)
+    - replace $a\geq 0 \perp b \geq$ by the equivalent expression $min(a,b)$ or the Fisher Burmeister function: $\sqrt(a^2+b^2) - (a+b)$ which has the same zeros (and is differentiable)
+
+
+
+----
+
 ## Time iteration
 
 Recall our optimality criterium:
@@ -122,6 +141,20 @@ Time iteration operator $T(\overrightarrow{x})$ is defined implicitly by:
 ## (Improved) Time iteration
 
 
+The recursive sequence $\overrightarrow{x_{n+1}} = T(\overrightarrow{x_n})$ converge at a geometric rate. Can we speed it up?
+
+Suppose we know how to compute $T^{\prime}$. Can we do a smart gradient descent from there?
+
+We look for a fixed point $$\overrightarrow{x}= T(\overrightarrow{x})$$
+
+We can do the Taylor expansion
+$ \overrightarrow{x} = T(x_n) + T^{\prime}(\overrightarrow{x_n})\left(\overrightarrow{x}  - \overrightarrow{x_{n+1}} \right)$
+
+Taking the difference: 
+$ \overrightarrow{x_{n+1}} - \overrightarrow{x} =  T^{\prime}(\overrightarrow{x_n})\left( \overrightarrow{x_{n}}- \overrightarrow{x}  \right)$
+
+A good guess for the steady-state:
+$$\overrightarrow{x}  = (I-T^{\prime})^{-1} \left(\overrightarrow{x_{n+1}} - T^{\prime}(\overrightarrow{x_{n}})\right)$$
 
 ----
